pyproject.toml
[project]
name = "requisitos-service"
version = "0.1.0"
description = "Sistema de Gestão de Requisitos de Software"
authors = [
    { name = "Seu Nome", email = "seuemail@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "python-dotenv",
    "pyproject-toml>=0.1.0",
    "passlib>=1.7.4",
    "bcrypt>=4.2.1",
    "python-jose>=3.3.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.workspace]
members = ["lock"]
